refactored misc queries into individual functions
